+++ tracking.py
@@ -81,14 +81,14 @@ def mouse_callback(event, x, y, flags, params):
     cv2.imshow('Select Markers', cur_frame)
     cv2.moveWindow('Select Markers', 50, 50)
 
-
-def select_markers(cap, bbox_size, frame_start, preprocessVals = None):
+def select_markers(cap, bbox_size, frame_start, preprocessing_vals = None):
     """event loop for handling initial marker selection
 
     Args:
         cap (cv2.VideoCapture): loaded video file for selecting markers of
         bbox_size (int): size of bounding box that is spec'd by user in ui
         frame_start (int): starting frame of video indicated in the frame selection tool
+        preprocessing_vals: the preprocessing values, if any
 
     Returns:
         mouse_params['marker_positions'] (list): selected marker positions
@@ -97,16 +97,29 @@ def select_markers(cap, bbox_size, frame_start, preprocessVals = None):
     cap.set(cv2.CAP_PROP_POS_FRAMES, frame_start)
     ret, first_frame = cap.read() # get first frame for selection
 
-    gray_frame = cv2.cvtColor(first_frame, cv2.COLOR_BGR2GRAY)
+    scaled_frame, scale_factor = scale_frame(first_frame)
 
-    if preprocessVals is not None:
-        print(preprocessVals)
-        first_frame = preprocess_frame(gray_frame, preprocessVals)
+    gray_frame = cv2.cvtColor(scaled_frame, cv2.COLOR_BGR2GRAY)
+
+    if preprocessing_vals is not None:
+        disp_preprocessing_vals = preprocessing_vals.copy()
+        disp_preprocessing_vals["Binarize"] = False
+
+        preprocessed_frame = preprocess_frame(gray_frame, disp_preprocessing_vals)
+
+        # print("Marker preprocessing done")
+
+        cv2.imshow('After Preprocessing', preprocessed_frame)
+        cv2.waitKey(0)  # Wait for a key press to proceed
+    else:
+        preprocessed_frame = gray_frame
     
-    cv2.imshow('Select Markers', first_frame) # show first frame
+    # print("Image show")
+    
+    cv2.imshow('Select Markers', preprocessed_frame) # show first frame
     cv2.moveWindow('Select Markers', 50, 50)
 
-    mouse_params = {"first_frame": first_frame.copy(), "marker_positions": [], 'bbox_size': bbox_size}
+    mouse_params = {"first_frame": preprocessed_frame.copy(), "marker_positions": [], 'bbox_size': bbox_size}
     cv2.setMouseCallback('Select Markers', mouse_callback, mouse_params) # set mouse callback function defn above
     
     # inf loop until user hits esc to cancel or enter to confirm selections
@@ -235,30 +248,49 @@ def init_trackers(marker_positions, bbox_size, first_frame, tracker_choice=Track
 
     return trackers
 
-def preprocess_frame(frame, preprocessVals):
+def preprocess_frame(frame, preprocessing_vals):
+    
+    # print("Preprocessing...")
     # Initialize a variable to store the modified frame
     modified_frame = frame.copy()
 
     # Apply improved smooth
-    if preprocessVals["Smoothness"] != 0:
-        modified_frame = improve_smoothing(modified_frame, preprocessVals["Smoothness"]/50+.1)
+    if preprocessing_vals["Smoothness"] != 0:
+        modified_frame = improve_smoothing(modified_frame, preprocessing_vals["Smoothness"]/50+.1)
 
     # Apply sharpening
-    if preprocessVals["Blur/Sharpness"] != 0:
-        modified_frame = sharpen_frame(modified_frame, preprocessVals["Blur/Sharpness"])
+    if preprocessing_vals["Blur/Sharpness"] != 0:
+        modified_frame = sharpen_frame(modified_frame, preprocessing_vals["Blur/Sharpness"])
 
     # Apply contrast enhancement
-    if preprocessVals["Contrast"] > 0:
-        modified_frame = enhance_contrast(modified_frame, preprocessVals["Contrast"])
+    if preprocessing_vals["Contrast"] > 0:
+        modified_frame = enhance_contrast(modified_frame, preprocessing_vals["Contrast"])
     
     # Apply brightness adjustment
-    if preprocessVals["Brightness"] != 0:
-        modified_frame = adjust_gamma(modified_frame, preprocessVals["Brightness"])
+    if preprocessing_vals["Brightness"] != 0:
+        modified_frame = adjust_gamma(modified_frame, preprocessing_vals["Brightness"])
 
-    if preprocessVals["Binarize"]:
+    if preprocessing_vals["Binarize"]:
         modified_frame = improve_binarization(modified_frame)
 
+    # intermediate_frame_check("Preprocessing done", modified_frame)
+    # print("Preprocessing done")
+
     return modified_frame
+
+def intermediate_frame_check(frame_name, frame):
+    """
+        Allows user to check a frame. Creates a new window displaying that frame, waits for keypress, then destroys itself
+
+    Args:
+        frame_name (string): desired name of the window
+        frame (frame): the frame to be displayed
+    """    
+    cv2.imshow(frame_name, frame)
+    cv2.moveWindow(frame_name, 50, 50)
+    cv2.waitKey(0)  # Wait for a key press to proceed
+    cv2.destroyWindow(frame_name)
+
 
 def enhance_contrast(frame, strength=50):
     # Define parameters for contrast enhancement
@@ -967,7 +999,7 @@ def track_area(
         video_file_name,
         data_label,
         preprocessing_need,
-        preprocessVals = None
+        preprocessing_vals = None
     ):
 
     """
@@ -1017,11 +1049,20 @@ def track_area(
         '1-data_label': data_label
     }
 
-    if preprocessVals is not None:
-        first_frame = preprocess_frame(first_frame, preprocessVals)
+    scaled_frame, scale_factor = scale_frame(first_frame)  # scale the frame
+
+    gray_frame = cv2.cvtColor(scaled_frame, cv2.COLOR_BGR2GRAY) # grayscale conversion
+
+    if preprocessing_vals is not None:
+        first_frame = preprocess_frame(gray_frame, preprocessing_vals)
+        disp_preprocessing_vals = preprocessing_vals.copy()
+        disp_preprocessing_vals["Binarize"] = False
+        print("Track area preprocessing done")
+    else:
+        first_frame = gray_frame
 
 
-    trackers = init_trackers(marker_positions, bbox_size, improve_smoothing(first_frame), TrackerChoice.CSRT)
+    trackers = init_trackers(marker_positions, bbox_size, first_frame, TrackerChoice.CSRT)
 
     while frame_num < frame_end:
         ret, frame = cap.read()
@@ -1037,8 +1078,10 @@ def track_area(
         gray_frame = cv2.cvtColor(scaled_frame, cv2.COLOR_BGR2GRAY)
 
         # Frame preprocessing
-        if preprocessVals is not None:
-            preprocessed_frame = preprocess_frame(gray_frame, preprocessVals)
+        if preprocessing_vals is not None:
+            preprocessed_frame = preprocess_frame(gray_frame, preprocessing_vals)
+            display_frame = preprocess_frame(gray_frame, disp_preprocessing_vals)
+            # print("In-loop area preprocessing done")
         else:
             preprocessed_frame = gray_frame
 
@@ -1088,9 +1131,9 @@ def track_area(
         # draw the chosen contour
         contour = contours[max_area_idx]
         if centroid:
-            cv2.circle(scaled_frame, centroid, 5, (0, 0, 255), -1)
-        cv2.drawContours(scaled_frame, [contour], -1, (255, 0, 0), 2)
-        cv2.putText(scaled_frame, str(i+1), (cx, cy), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (255, 255, 255), 1)
+            cv2.circle(display_frame, centroid, 5, (0, 0, 255), -1)
+        cv2.drawContours(display_frame, [contour], -1, (255, 0, 0), 2)
+        cv2.putText(display_frame, str(i+1), (cx, cy), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (255, 255, 255), 1)
 
         # record data
         print(centroid)
@@ -1106,10 +1149,10 @@ def track_area(
         except TypeError:
             print("centroid not found")
 
-
-        #cv2.imshow('Surface Area Tracking', noise_reduced_frame)
-        cv2.imshow('Surface Area Tracking', scaled_frame)
-
+        cv2.imshow('Surface Area Tracking', display_frame)
+        # cv2.imshow('Surface Area Tracking', preprocessed_frame)
+        # cv2.imshow('Surface Area Tracking', gray_frame)
+        
         if cv2.waitKey(1) == 27:
             break